pyproject.toml
[project]
name = "rfm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud>=0.34.0",
    "google-cloud-bigquery>=3.35.0",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "polars>=1.31.0",
    "scikit-learn>=1.7.1",
    "streamlit>=1.47.0",
]
